requirements.txt
asgiref==3.7.2
attrs==23.2.0
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.3
django-cache-url==3.4.5
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.1
environs==11.0.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.21.1
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
whitenoise==6.6.0
